Rename VuFind\Mailer\Factory to MailerFactory for consistency.
