+++ module/VuFind/tests/unit-tests/src/VuFindTest/Mailer/MailerTest.php
@@ -31,7 +31,7 @@ namespace VuFindTest\Mailer;
 
 use Symfony\Component\Mailer\MailerInterface;
 use Symfony\Component\Mime\Address;
-use VuFind\Mailer\Factory as MailerFactory;
+use VuFind\Mailer\MailerFactory;
 use VuFind\Mailer\Mailer;
 use VuFindTest\Container\MockContainer;
